feat(pstor/watcher): add pstor watcher
Adds a pstor watcher api which registers watches with the pstor(etcd) and
also ensures the watches are live, reconnecting the stream if necessary.
A single stream per watch is used at the moment.
This might not scale well if we need to have a high watch update rate
though we don't intend it to be so.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>